Fixed various warnings
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@1208 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Apps/Subdivider/SubdivideWidget.cc
@@ -268,9 +268,9 @@ void SubdivideWidget::keyPressEvent( QKeyEvent *k )
 
 void SubdivideWidget::update()
 {
-  unsigned int n_faces = viewer_widget_->mesh().n_faces();
-  unsigned int n_edges = viewer_widget_->mesh().n_edges();
-  unsigned int n_vertices = viewer_widget_->mesh().n_vertices();
+  size_t n_faces = viewer_widget_->mesh().n_faces();
+  size_t n_edges = viewer_widget_->mesh().n_edges();
+  size_t n_vertices = viewer_widget_->mesh().n_vertices();
   QString message(""), temp;
   message.append(temp.setNum(n_faces));
   message.append(" Faces, ");
